reorganize code add test fro entropy
