+++ pkg/scan/post_scan.go
@@ -2,16 +2,13 @@ package scan
 
 import (
 	_ "embed" // Embed kics CLI img and scan-flags
-	"encoding/json"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 	"time"
 
 	consoleHelpers "github.com/Checkmarx/kics/internal/console/helpers"
 	"github.com/Checkmarx/kics/pkg/engine/provider"
-	secrets "github.com/Checkmarx/kics/pkg/engine/secrets"
 	"github.com/Checkmarx/kics/pkg/model"
 	consolePrinter "github.com/Checkmarx/kics/pkg/printer"
 	"github.com/Checkmarx/kics/pkg/progress"
@@ -144,140 +141,4 @@ func (c *Client) postScan(scanResults *Results) error {
 	}
 
 	return nil
-}
-
-func maskPreviewLines(secretsPath string, scanResults *Results) error {
-	secretsRegexRulesContent, err := getSecretsRegexRules(secretsPath)
-	if err != nil {
-		return err
-	}
-
-	var allRegexQueries secrets.RegexRuleStruct
-
-	err = json.Unmarshal([]byte(secretsRegexRulesContent), &allRegexQueries)
-	if err != nil {
-		return err
-	}
-
-	allowRules, err := secrets.CompileRegex(allRegexQueries.AllowRules)
-	if err != nil {
-		return err
-	}
-
-	rules, err := compileRegexQueries(allRegexQueries.Rules)
-	if err != nil {
-		return err
-	}
-
-	for i := range scanResults.Results {
-		item := scanResults.Results[i]
-		hideSecret(item.VulnLines, &allowRules, &rules)
-	}
-	return nil
-}
-
-func compileRegexQueries(allRegexQueries []secrets.RegexQuery) ([]secrets.RegexQuery, error) {
-
-	for i := range allRegexQueries {
-		compiledRegexp, err := regexp.Compile(allRegexQueries[i].RegexStr)
-		if err != nil {
-			return allRegexQueries, err
-		}
-		allRegexQueries[i].Regex = compiledRegexp
-		//TODO: check if you can remove these lines.
-		for j := range allRegexQueries[i].AllowRules {
-			allRegexQueries[i].AllowRules[j].Regex = regexp.MustCompile(allRegexQueries[i].AllowRules[j].RegexStr)
-		}
-	}
-	return allRegexQueries, nil
-}
-
-func hideSecret(lines *[]model.CodeLine, allowRules *[]secrets.AllowRule, rules *[]secrets.RegexQuery) {
-	for idx, line := range *lines {
-		for i := range *rules {
-			rule := (*rules)[i]
-			isSecret, groups := isSecret(line.Line, &rule, allowRules)
-			// if not a secret skip to next line
-			if !isSecret {
-				continue
-			}
-
-			if rule.SpecialMask == "all" {
-				(*lines)[idx].Line = "<SECRET-MASKED-ON-PURPOSE>"
-				continue
-			}
-
-			if len(rule.Entropies) == 0 {
-				maskSecret(&rule, lines, idx)
-			}
-
-			//TODO: check if you should remove this entire segment
-			for i := range rule.Entropies {
-				entropy := rule.Entropies[i]
-
-				// if matched group does not exist continue
-				if len(groups[0]) <= entropy.Group {
-					return
-				}
-
-				//TODO: check if we need to mask. when does this happen?
-				maskSecret(&rule, lines, idx)
-			}
-		}
-	}
-}
-
-func maskSecret(rule *secrets.RegexQuery, lines *[]model.CodeLine, idx int) {
-	regex := rule.RegexStr
-	line := (*lines)[idx]
-
-	if len(rule.SpecialMask) > 0 {
-		regex = "(.+)" + rule.SpecialMask
-	}
-
-	var re = regexp.MustCompile(regex)
-	match := re.FindString(line.Line)
-
-	if len(rule.SpecialMask) > 0 {
-		match = line.Line[len(match):]
-	}
-
-	if match != "" {
-		(*lines)[idx].Line = strings.Replace(line.Line, match, "<SECRET-MASKED-ON-PURPOSE>", 1)
-	} else {
-		(*lines)[idx].Line = "<SECRET-MASKED-ON-PURPOSE>"
-	}
-}
-
-func isSecret(line string, rule *secrets.RegexQuery, allowRules *[]secrets.AllowRule) (isSecretRet bool, groups [][]string) {
-	if secrets.IsAllowRule(line, *allowRules) {
-		return false, [][]string{}
-	}
-
-	groups = rule.Regex.FindAllStringSubmatch(line, -1)
-
-	for _, group := range groups {
-		splitedText := strings.Split(line, "\n")
-		max := -1
-		for i, splited := range splitedText {
-			if len(groups) < rule.Multiline.DetectLineGroup {
-				if strings.Contains(splited, group[rule.Multiline.DetectLineGroup]) && i > max {
-					max = i
-				}
-			}
-		}
-		if max == -1 {
-			continue
-		}
-		secret, newGroups := isSecret(strings.Join(append(splitedText[:max], splitedText[max+1:]...), "\n"), rule, allowRules)
-		if !secret {
-			continue
-		}
-		groups = append(groups, newGroups...)
-	}
-
-	if len(groups) > 0 {
-		return true, groups
-	}
-	return false, [][]string{}
 }
